chore(clippy): clear all default-lint warnings across workspace
Investigated each warning for root cause; none indicated missing code,
defects, or bad merges. Fixes split by file:

ropus/src/lib.rs
- Consolidate silk_trace module docs into a single inner //! block
  (was emitting clippy::mixed_attributes_style).

ropus/src/silk/decoder.rs
- Replace `nb_subfr.saturating_sub(2).max(0)` with plain `nb_subfr - 2`
  in silk_plc_update; nb_subfr is usize so .max(0) was unreachable, and
  the next line already uses `nb_subfr - 1` with the same precondition.
  Matches reference C `silk/PLC.c:186` exactly.

ropus-tools-core/src/commands/encode.rs
- Use `checked_div` for the avg-bitrate computation rather than a
  manual `if samples_written > 0` guard (clippy::manual_checked_ops,
  new in Rust 1.95).

harness/src/bin_inner/projection_roundtrip.rs
- Drop a no-op `as i32` on c_int (i32 on Linux).

harness/src/bin_inner/fuzz_repro_diff.rs
- Reword module-doc "+" so it isn't parsed as a markdown list bullet.
- Replace three `(x - k) % n != 0` patterns with `!(x-k).is_multiple_of(n)`.
- Add `BaselineSetter` type alias for the function-pointer tuple
  (clippy::type_complexity).
- Drop a no-op `as i32` on an already-i32 expression.
- Use struct-literal init with `..Default::default()` to silence
  field_reassign_with_default in dump_rust_extended.

Verified clean: cargo build, cargo fmt --check, cargo clippy
--workspace --all-targets, plus clippy with --features
trace-silk-encode. Targeted tests still pass (silk::decoder, opus::
encoder, ropus-tools-core).

+++ harness/src/bin_inner/projection_roundtrip.rs
@@ -318,7 +318,7 @@ fn run_fixture(path: &Path, bitrate: i32) -> FixtureResult {
         );
     }
     assert_eq!(
-        r_gain, c_enc.demixing_matrix_gain as i32,
+        r_gain, c_enc.demixing_matrix_gain,
         "demixing matrix gain mismatch: ropus={}, C={}",
         r_gain, c_enc.demixing_matrix_gain
     );
